fix: clean up terminating logic in trial.go
This commit fixes a deadlock in trial actor. It happens
when a trial is terminated gracefully and the trial containers
are still starting. Before all the trial containers are
started, the trial would not send the rendezvous information,
which blocks trial containers sending back completed workload
messages. This issue exists on both the default and k8s resource
providers. In the default resource provider only when the instances
are launched. However, on k8s, containers are started one by one.
There might be a situation that some containers would never start
up. This commit adds a timeout for all graceful trial terminating
and directly handles as killing if not all containers are connected.

The solution is to check if all containers are ready when gracefully
terminating it. If not, rather than waiting for the completed workloads,
kill all the containers. Also, add a timeout to all places that
gracefully terminate the trial.